# challenge/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Qualifica",
        "description": "Está é a documentdação da api do desafio",
        "contact": {
            "email": "lincolnalca@gmail.com"
        },
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/user/{id}/downgrade": {
            "put": {
                "tags": [
                    "User"
                ],
                "operationId": "53b2b8aa5b28de5025c91d3d12879c94",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Store successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "413": {
                        "description": "Store failed",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    }
}